Filter data if user has no pattern
